SFDM-3469 [back][product] Разработать метод для агрегации результатов ФФ

+++ src/main/java/ru/beeline/fdmproducts/dto/DomainDTO.java
@@ -10,9 +10,9 @@ import java.util.List;
 @Builder
 @ToString
 public class DomainDTO {
-    private Long id;
+    private Integer id;
     private String name;
     private String alias;
-    private Long ownerId;
+    private Integer ownerId;
     private List<FitnessFunctionProductDTO> product;
 }
